Split core package into core and reco
The modules moved from core to reco are

Closes #150

+ `nh5`
+ `params`
+ `peak_functions`
+ `pmaps_functions`
+ `tbl_functions`
+ `wfm_functions`
+ `wfm_to_df_functions`
