+++ invisible_cities/cities/isidora.py
@@ -6,8 +6,8 @@ from time import time
 import numpy as np
 import tables as tb
 
-from   invisible_cities.core.nh5 import RunInfo, EventInfo
-import invisible_cities.core.tbl_functions as tbl
+from   invisible_cities.reco.nh5 import RunInfo, EventInfo
+import invisible_cities.reco.tbl_functions as tbl
 from   invisible_cities.core.configure import configure
 from   invisible_cities.core.system_of_units_c import units
 from   invisible_cities.cities.base_cities import (DeconvolutionCity,
