+++ invisible_cities/cities/maurilia.py
@@ -15,9 +15,9 @@ from time import time
 import numpy  as np
 import tables as tb
 
-import invisible_cities.core.tbl_functions as tbl
+import invisible_cities.reco.tbl_functions as tbl
 from   invisible_cities.core.configure import configure, print_configuration
-from   invisible_cities.core.nh5 import MCTrack
+from   invisible_cities.reco.nh5 import MCTrack
 from   invisible_cities.cities.base_cities import City
 
 class Maurilia(City):
