+++ invisible_cities/core/core_functions_performance.py
@@ -5,8 +5,8 @@ Performance of some coreFunctions
 
 from time import time
 import numpy as np
-import invisible_cities.core.wfm_functions as wfm
-import invisible_cities.core.peak_functions_c as cpf
+import invisible_cities.reco.wfm_functions as wfm
+import invisible_cities.reco.peak_functions_c as cpf
 
 
 t = np.arange(1.,100., 0.1, dtype=np.double)
@@ -28,4 +28,4 @@ t0 = time()
 T, E = cpf.rebin_waveform(t, e, stride=2)
 t1 = time()
 dt = t1 - t0
-print("cython rebin_waveform (ICython/Core/peakFunctions) run in {} s".format(dt))
+print("cython rebin_waveform (Reco/peakFunctions) run in {} s".format(dt))
